+++ ThirdPartyAdapters/dtexchange/dtexchange/src/test/kotlin/com/google/ads/mediation/fyber/DTExchangeBannerAdTest.kt
@@ -10,6 +10,7 @@ import com.fyber.inneractive.sdk.external.InneractiveAdViewUnitController
 import com.fyber.inneractive.sdk.external.InneractiveErrorCode
 import com.google.ads.mediation.adaptertestkit.AdErrorMatcher
 import com.google.ads.mediation.adaptertestkit.AdapterTestKitConstants.TEST_BID_RESPONSE
+import com.google.ads.mediation.adaptertestkit.AdapterTestKitConstants.TEST_WATERMARK
 import com.google.ads.mediation.adaptertestkit.createMediationBannerAdConfiguration
 import com.google.android.gms.ads.AdError
 import com.google.android.gms.ads.mediation.MediationAdLoadCallback
@@ -45,7 +46,13 @@ class DTExchangeBannerAdTest {
 
   @Before
   fun setUp() {
-    dtExchangeBannerAd = DTExchangeBannerAd(mockAdLoadCallback)
+    val adConfiguration =
+      createMediationBannerAdConfiguration(
+        context = context,
+        bidResponse = TEST_BID_RESPONSE,
+        watermark = TEST_WATERMARK,
+      )
+    dtExchangeBannerAd = DTExchangeBannerAd(adConfiguration, mockAdLoadCallback)
   }
 
   @Test